+++ examples/invokeCmd__fail.py
@@ -8,12 +8,10 @@ import jk_simpleexec
 
 
 
-cmdResult = jk_simpleexec.invokeCmd1(
+cmdResult = jk_simpleexec.invokeCmd2(
 	cmdPath = "/foo/bar/foobar",
 	cmdArgs = [
 	],
-	bRemoveTrailingNewLinesFromStdOut = True,			# this is the default
-	bRemoveTrailingNewLinesFromStdErr = True,			# this is the default
 	dataToPipeAsStdIn = None,							# this is the default
 	workingDirectory = "/",
 )