+++ src/jk_simpleexec/__init__.py
@@ -1,6 +1,6 @@
 ﻿
 
-__version__ = "0.2021.12.4"
+__version__ = "0.2021.12.14"
 
 
 
@@ -8,5 +8,7 @@ from .CommandResult import CommandResult
 from .TextDataProcessingPolicy import TextDataProcessingPolicy
 from ._DebugValveToFile import _DebugValveToFile
 from ._common import enableDebugging, DEFAULT_STDOUT_PROCESSING, DEFAULT_STDERR_PROCESSING, processCmdOutput
-from .simpleexec import invokeCmd, invokeCmd1
+from .simpleexec import invokeCmd, invokeCmd1, invokeCmd2
 from .invoke_utils import runCmd
+
+