+++ src/jk_simpleexec/simpleexec.py
@@ -22,7 +22,7 @@ from . import _common as _common
 # NOTE: Despite for <c>cmdPath</c> and <c>cmdArgs</c> do <b>not</b> rely on the order of the arguments. If you need to specify them invoke them by name as they are
 # intended to be invoked that way! Their order might be changed unnoticed in other versions of this API.
 #
-# NOTE: This method is deprecated and should no longer be called. Please use the new <c>invokeCmd1()</c> instead.
+# NOTE: This method is deprecated and should no longer be called. Please use the new <c>invokeCmd2()</c> instead.
 #
 # @param		string cmdPath						(required) The (absolute) path to the program to invoke.
 # @param		string[] cmdArgs					(required) A list of arguments. Specify <c>None</c> if you do not want to have any arguments.
@@ -126,6 +126,8 @@ def invokeCmd(
 #
 # Synchroneously invokes the specified command on the local machine. Output of STDOUT and STDERR is collected and returned by the <c>CommandResult</c> return object.
 #
+# NOTE: This method is deprecated and should no longer be called. Please use the new <c>invokeCmd2()</c> instead.
+#
 # @param		string cmdPath								(required) The (absolute) path to the program to invoke.
 # @param		string[] cmdArgs							(required) A list of arguments. Specify <c>None</c> if you do not want to have any arguments.
 #															Please note that there is no shell to interprete these commands.
@@ -152,6 +154,54 @@ def invokeCmd1(
 		stdErrProcessing:TextDataProcessingPolicy = None,
 		log = None,
 	) -> CommandResult:
+	return invokeCmd2(
+		cmdPath=cmdPath,
+		cmdArgs=cmdArgs,
+		dataToPipeAsStdIn=dataToPipeAsStdIn,
+		workingDirectory=workingDirectory,
+		stdOutProcessing=stdOutProcessing,
+		stdErrProcessing=stdErrProcessing,
+		shell=False,
+		log=log,
+	)
+#
+
+#
+# Synchroneously invokes the specified command on the local machine. Output of STDOUT and STDERR is collected and returned by the <c>CommandResult</c> return object.
+#
+# @param		string cmdPath								(required) The (absolute) path to the program to invoke.
+# @param		string[] cmdArgs							(required) A list of arguments. Specify <c>None</c> if you do not want to have any arguments.
+#															Please note that there is no shell to interprete these commands.
+# @param		str|bytes[] dataToPipeAsStdIn				(optional) Either a string or binary data (or None) that should be passed on to the application invoked usint STDIN.
+#															If string data is presented it is automatically encoded using UTF-8
+# @param		str workingDirectory						(optional) If you specify a working directory here this function will change to this working directory
+#															specified in <c>workingDirector</c> and return to the previous one after the command has been completed.
+# @param		TextDataProcessingPolicy stdOutProcessing	(optional) If specified you can override defaults of the STDOUT preprocessing that can already be done by this function.
+# @param		TextDataProcessingPolicy stdErrProcessing	(optional) If specified you can override defaults of the STDERR preprocessing that can already be done by this function.
+# @param		* log										(optional) You can specify a logger here. This logger will receive a notice about what command is going to be executed.
+#															For this to work the logger object specified here is checked against the following criterias (in the order listed):
+#															* have a method named `notice(..)` expecting a single string argument - the log message
+#															* have a method named `info(..)` expecting a single string argument - the log message
+#															* is callable (= is a method itself) expecting a single string argument - the log message
+# @param		bool shell									If set to `True` interpret the specified command by a shell. (This is then equivalent to `subprocess.Popen(..)`
+#															with `shell = True`.)
+#
+# @return		CommandOutput								Returns an object that contains the exit status, (preprocessed) STDOUT and (preprocessed) STDERR data.
+#
+def invokeCmd2(
+		*argv,
+		cmdPath:str,
+		cmdArgs:list,
+		dataToPipeAsStdIn:typing.Union[str,bytes,bytearray] = None,
+		workingDirectory:str = None,
+		stdOutProcessing:TextDataProcessingPolicy = None,
+		stdErrProcessing:TextDataProcessingPolicy = None,
+		shell:bool = False,
+		log = None,
+	) -> CommandResult:
+
+	if len(argv) > 0:
+		raise Exception("For compatibility with future changes please invoke this method with named arguments only!")
 
 	stdOutProcessing = _common.DEFAULT_STDOUT_PROCESSING.override(stdOutProcessing)
 	stdErrProcessing = _common.DEFAULT_STDERR_PROCESSING.override(stdErrProcessing)
@@ -208,10 +258,10 @@ def invokeCmd1(
 		# run the processes
 
 		if dataToPipeAsStdIn:
-			p = subprocess.Popen(cmd, shell=False, stdout=subprocess.PIPE, stderr=subprocess.PIPE, stdin=subprocess.PIPE)
+			p = subprocess.Popen(cmd, shell=shell, stdout=subprocess.PIPE, stderr=subprocess.PIPE, stdin=subprocess.PIPE)
 			p.stdin.write(dataToPipeAsStdIn)
 		else:
-			p = subprocess.Popen(cmd, shell=False, stdout=subprocess.PIPE, stderr=subprocess.PIPE, stdin=None)
+			p = subprocess.Popen(cmd, shell=shell, stdout=subprocess.PIPE, stderr=subprocess.PIPE, stdin=None)
 		(stdout, stderr) = p.communicate()
 
 		# process stdout